refactor(client/main): further organization
