+++ client/fleeca.lua
@@ -1,32 +1,12 @@
 local config = require 'config.client'
 local sharedConfig = require 'config.shared'
-isLoggedIn = LocalPlayer.state['isLoggedIn']
 currentThermiteGate = 0
 CurrentCops = 0
-isDrilling = false
 local closestBank = 0
 local inElectronickitZone = false
 local copsCalled = false
 local refreshed = false
 local currentLocker = 0
-
-AddEventHandler('onResourceStop', function(resource)
-    if resource ~= cache.resource then return end
-    ResetBankDoors()
-end)
-
-RegisterNetEvent('QBCore:Client:OnPlayerLoaded', function()
-    local config1, config2, config3 = lib.callback.await('qb-bankrobbery:server:GetConfig', false)
-    sharedConfig.powerStations = config1
-    sharedConfig.bigBanks = config2
-    sharedConfig.smallBanks = config3
-    ResetBankDoors()
-    isLoggedIn = true
-end)
-
-RegisterNetEvent('QBCore:Client:OnPlayerUnload', function()
-    isLoggedIn = false
-end)
 
 --- This is triggered once the hack at a small bank is done
 --- @param success boolean
@@ -36,151 +16,6 @@ local function onHackDone(success)
     if not success then return end
     TriggerServerEvent('qb-bankrobbery:server:setBankState', closestBank)
 end
-
---- This will trigger the sequence of opening a safe locker of a bank
---- @param bankId string | number
---- @param lockerId number
---- @return nil
-function openLocker(bankId, lockerId) -- Globally Used
-    local pos = GetEntityCoords(cache.ped)
-    DropFingerprint()
-    TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isBusy', true)
-    if bankId == 'paleto' then
-        local hasItem = exports.ox_inventory:Search('count', 'drill') > 0
-        if hasItem then
-            -- loadAnimDict('anim@heists@fleeca_bank@drilling')
-            -- TaskPlayAnim(cache.ped, 'anim@heists@fleeca_bank@drilling', 'drill_straight_idle', 3.0, 3.0, -1, 1, 0, false, false, false)
-            local drillObject = CreateObject(`hei_prop_heist_drill`, pos.x, pos.y, pos.z, true, true, true)
-            AttachEntityToEntity(drillObject, cache.ped, GetPedBoneIndex(cache.ped, 57005), 0.14, 0, -0.01, 90.0, -90.0, 180.0, true, true, false, true, 1, true)
-            isDrilling = true
-            if lib.progressBar({
-                duration = 25000,
-                label = Lang:t('general.breaking_open_safe'),
-                canCancel = true,
-                useWhileDead = false,
-                disable = {
-                    move = true,
-                    car = true,
-                    mouse = false,
-                    combat = true
-                },
-                anim = {
-                    dict = 'anim@heists@fleeca_bank@drilling',
-                    clip = 'drill_straight_idle',
-                    flag = 1
-                }
-            }) then
-                DetachEntity(drillObject, true, true)
-                DeleteObject(drillObject)
-                TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isOpened', true)
-                TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isBusy', false)
-                TriggerServerEvent('qb-bankrobbery:server:recieveItem', 'paleto', bankId, lockerId)
-                exports.qbx_core:Notify(Lang:t('success.success_message'), 'success')
-                SetTimeout(500, function()
-                    isDrilling = false
-                end)
-            else
-                TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isBusy', false)
-                DetachEntity(drillObject, true, true)
-                DeleteObject(drillObject)
-                exports.qbx_core:Notify(Lang:t('error.cancel_message'), 'error')
-                SetTimeout(500, function()
-                    isDrilling = false
-                end)
-            end
-        else
-            exports.qbx_core:Notify(Lang:t('error.safe_too_strong'), 'error')
-            TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isBusy', false)
-        end
-    elseif bankId == 'pacific' then
-        local hasItem = exports.ox_inventory:Search('count', 'drill') > 0
-        if hasItem then
-            local drillObject = CreateObject(`hei_prop_heist_drill`, pos.x, pos.y, pos.z, true, true, true)
-            AttachEntityToEntity(drillObject, cache.ped, GetPedBoneIndex(cache.ped, 57005), 0.14, 0, -0.01, 90.0, -90.0, 180.0, true, true, false, true, 1, true)
-            isDrilling = true
-            if lib.progressBar({
-                duration = 25000,
-                label = Lang:t('general.breaking_open_safe'),
-                canCancel = true,
-                useWhileDead = false,
-                disable = {
-                    move = true,
-                    car = true,
-                    mouse = false,
-                    combat = true
-                },
-                anim = {
-                    dict = 'anim@heists@fleeca_bank@drilling',
-                    clip = 'drill_straight_idle',
-                    flag = 1
-                }
-            }) then
-                DetachEntity(drillObject, true, true)
-                DeleteObject(drillObject)
-
-                TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isOpened', true)
-                TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isBusy', false)
-                TriggerServerEvent('qb-bankrobbery:server:recieveItem', 'pacific', bankId, lockerId)
-                exports.qbx_core:Notify(Lang:t('success.success_message'), 'success')
-                SetTimeout(500, function()
-                    isDrilling = false
-                end)
-            else
-                TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isBusy', false)
-                DetachEntity(drillObject, true, true)
-                DeleteObject(drillObject)
-                exports.qbx_core:Notify(Lang:t('error.cancel_message'), 'error')
-                SetTimeout(500, function()
-                    isDrilling = false
-                end)
-            end
-        else
-            exports.qbx_core:Notify(Lang:t('error.safe_too_strong'), 'error')
-            TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isBusy', false)
-        end
-    else
-        isDrilling = true
-        if lib.progressBar({
-            duration = 32000,
-            label = Lang:t('general.breaking_open_safe'),
-            canCancel = true,
-            useWhileDead = false,
-            disable = {
-                move = true,
-                car = true,
-                mouse = false,
-                combat = true
-            },
-            anim = {
-                dict = 'anim@gangops@facility@servers@',
-                clip = 'hotwire',
-                flag = 1
-            }
-        }) then
-            TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isOpened', true)
-            TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isBusy', false)
-            TriggerServerEvent('qb-bankrobbery:server:recieveItem', 'small', bankId, lockerId)
-            exports.qbx_core:Notify(Lang:t('success.success_message'), 'success')
-            SetTimeout(500, function()
-                isDrilling = false
-            end)
-        else
-            TriggerServerEvent('qb-bankrobbery:server:setLockerState', bankId, lockerId, 'isBusy', false)
-            exports.qbx_core:Notify(Lang:t('error.cancel_message'), 'error')
-            SetTimeout(500, function()
-                isDrilling = false
-            end)
-        end
-    end
-    CreateThread(function()
-        while isDrilling do
-            TriggerServerEvent('hud:server:GainStress', math.random(4, 8))
-            Wait(10000)
-        end
-    end)
-end
-
--- Events
 
 RegisterNetEvent('electronickit:UseElectronickit', function()
     DropFingerprint()
@@ -225,19 +60,6 @@ RegisterNetEvent('electronickit:UseElectronickit', function()
     end
 end)
 
-RegisterNetEvent('qb-bankrobbery:client:setBankState', function(bankId)
-    if bankId == 'paleto' then
-        sharedConfig.bigBanks.paleto.isOpened = true
-        OpenPaletoDoor()
-    elseif bankId == 'pacific' then
-        sharedConfig.bigBanks.pacific.isOpened = true
-        OpenPacificDoor()
-    else
-        sharedConfig.smallBanks[bankId].isOpened = true
-        OpenFleecaDoor(bankId)
-    end
-end)
-
 RegisterNetEvent('qb-bankrobbery:client:enableAllBankSecurity', function()
     for k in pairs(sharedConfig.smallBanks) do
         sharedConfig.smallBanks[k].alarm = true
@@ -262,70 +84,11 @@ RegisterNetEvent('qb-bankrobbery:client:BankSecurity', function(key, status)
     end
 end)
 
-RegisterNetEvent('qb-bankrobbery:client:setLockerState', function(bankId, lockerId, state, bool)
-    if bankId == 'paleto' then
-        sharedConfig.bigBanks.paleto.lockers[lockerId][state] = bool
-    elseif bankId == 'pacific' then
-        sharedConfig.bigBanks.pacific.lockers[lockerId][state] = bool
-    else
-        sharedConfig.smallBanks[bankId].lockers[lockerId][state] = bool
-    end
-end)
-
 RegisterNetEvent('qb-bankrobbery:client:ResetFleecaLockers', function(BankId)
     sharedConfig.smallBanks[BankId].isOpened = false
     for k in pairs(sharedConfig.smallBanks[BankId].lockers) do
         sharedConfig.smallBanks[BankId].lockers[k].isOpened = false
         sharedConfig.smallBanks[BankId].lockers[k].isBusy = false
-    end
-end)
-
-RegisterNetEvent('qb-bankrobbery:client:robberyCall', function(type, coords)
-    if not isLoggedIn then return end
-    local PlayerJob = exports.qbx_core:GetPlayerData().job
-    if PlayerJob.name ~= 'police' or not PlayerJob.onduty then return end
-    if type == 'small' then
-        PlaySound(-1, 'Lose_1st', 'GTAO_FM_Events_Soundset', false, 0, true)
-        TriggerServerEvent('police:server:policeAlert', Lang:t('general.fleeca_robbery_alert'))
-    elseif type == 'paleto' then
-        PlaySound(-1, 'Lose_1st', 'GTAO_FM_Events_Soundset', false, 0, true)
-        Wait(100)
-        PlaySoundFrontend( -1, 'Beep_Red', 'DLC_HEIST_HACKING_SNAKE_SOUNDS', true)
-        Wait(100)
-        PlaySound(-1, 'Lose_1st', 'GTAO_FM_Events_Soundset', false, 0, true)
-        Wait(100)
-        PlaySoundFrontend( -1, 'Beep_Red', 'DLC_HEIST_HACKING_SNAKE_SOUNDS', true)
-        TriggerServerEvent('police:server:policeAlert', Lang:t('general.paleto_robbery_alert'))
-    elseif type == 'pacific' then
-        PlaySound(-1, 'Lose_1st', 'GTAO_FM_Events_Soundset', false, 0, true)
-        Wait(100)
-        PlaySoundFrontend( -1, 'Beep_Red', 'DLC_HEIST_HACKING_SNAKE_SOUNDS', true)
-        Wait(100)
-        PlaySound(-1, 'Lose_1st', 'GTAO_FM_Events_Soundset', false, 0, true)
-        Wait(100)
-        PlaySoundFrontend( -1, 'Beep_Red', 'DLC_HEIST_HACKING_SNAKE_SOUNDS', true)
-        TriggerServerEvent('police:server:policeAlert', Lang:t('general.pacific_robbery_alert'))
-    end
-    local transG = 250
-    local blip = AddBlipForCoord(coords.x, coords.y, coords.z)
-    SetBlipSprite(blip, 487)
-    SetBlipColour(blip, 4)
-    SetBlipDisplay(blip, 4)
-    SetBlipAlpha(blip, transG)
-    SetBlipScale(blip, 1.2)
-    SetBlipFlashes(blip, true)
-    BeginTextCommandSetBlipName('STRING')
-    AddTextComponentString(Lang:t('general.bank_robbery_police_call'))
-    EndTextCommandSetBlipName(blip)
-    while transG ~= 0 do
-        Wait(180 * 4)
-        transG = transG - 1
-        SetBlipAlpha(blip, transG)
-        if transG == 0 then
-            SetBlipSprite(blip, 2)
-            RemoveBlip(blip)
-            return
-        end
     end
 end)
